Let bors cut after <details>
I did not think about dependabot when adding the cut after "---".
Dependabot adds HTML in the PR title and we don't want that to be in the
commit message for the merge commit, do we?

So change it to "<details>" here, which is less ergonomic for users, but
way better for merging dependabot updates.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>