Implement handling of received Puback
We introduce a oneshot channel here because we need to re-unite the
read- and write-half if we need to reconnect.

We need to re-unite them because we want to empty the read-half before
performing the reconnect, so that we do not accidentially use a
the connection state from before the reconnect, after we reconnected.

Co-authored-by: Marcel Müller <neikos@neikos.email>
Signed-off-by: Marcel Müller <neikos@neikos.email>
Signed-off-by: Matthias Beyer <mail@beyermatthias.de>