Rename handle functions
We call these "handle" functions only after receiving information from
the server (for now), so there's no decision-process to be implemented
by the user.
There's no need to be async here, and the user cannot return values to
cloudmqtt in this case.
Therefore, we call these functions `on_*` because we simply notify the
user of something that happened.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>